+++ doc/source/CHANGELOG.rst
@@ -9,6 +9,8 @@ Version 1.7.0:
  - add dedicated CSV/TSV output (:issue:`10`)
  - fix loading reference database from running ARB (:issue:`76`)
  - report errors when sequence can't be read from ARB (:issue:`73`)
+ - add :option:`--arb-list-fields` listing fields available in ARB
+   database
 
 Version 1.6.1:
 --------------